Fix scalability issues with malloc_size implementation
The malloc_size tcmalloc implementation use the MallocExtension class
to call GetAllocatedSize to return the object size. The issue with is
the MallocExtension::instance with uses a multithread singleton pattern
to initialize the default MallocExtension. This pattern use the
perftools_pthread_once, and where it garantees to work, it may create
scalability issues if the call is called frequenty (like the redis
database program).

This patch changes the way the default obejct is created to being
static, trading space by performance.